+++ ChangeLog.md
@@ -8,6 +8,7 @@
 - Add time units to slot duration and epoch duration in consensus status.
 - Update `register-data` command to register data as CBOR encoded strings or JSON using the new flags
   `--string` and `--json`. Raw data can still be registered using the new flag `--raw`.
+- Add `raw DisconnectPeer`, a dual to the existing `raw ConnectPeer`.
 
 ## 1.1.1