+++ ChangeLog.md
@@ -9,6 +9,10 @@
 - Update `register-data` command to register data as CBOR encoded strings or JSON using the new flags
   `--string` and `--json`. Raw data can still be registered using the new flag `--raw`.
 - Add `raw DisconnectPeer`, a dual to the existing `raw ConnectPeer`.
+- Warn user when trying to add a baker with stake below the minimum threshold.
+- Improve how contract schemas are shown as JSON.
+  - Display complex types in arrays correctly.
+  - Use angle brackets to indicate placeholders, e.g. `"<UInt16>"` instead of `"UInt16"`.
 - Improve `module inspect` so it shows all contracts from a module regardless of
 whether a schema is included or not.